[10.0][MIG][REF] Splitted file to separate objects sale.order.line from sale.order .
